just update

+++ app/Http/Controllers/AdminController.php
@@ -17,8 +17,9 @@ use Illuminate\Support\Facades\File;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
+use Illuminate\Support\Facades\URL;
 use Illuminate\Support\Facades\Validator;
-
+use Intervention\Image\ImageManagerStatic as Image;
 class AdminController extends Controller
 {
 /**
@@ -132,7 +133,9 @@ class AdminController extends Controller
 
             case 'tour':
                 $obj = Tour::find($request->id);
-                File::deleteDirectory('images/tours/'.$obj->id);
+                if(File::exists('images/tours/'.$obj->id)) {
+                    File::deleteDirectory('images/tours/'.$obj->id);
+                }
                 if(null == $obj){
                     $status = false;
                     break;
@@ -215,7 +218,26 @@ class AdminController extends Controller
 
     public function postUpdateCropped(Request $request)
     {
-        dd($request);
+//        dd($request);
+        $image = $request->file('data');
+        $source = str_replace(URL::to('/') . '/', '', $request->source);
+        $image_name = array_reverse(explode('/',$source))[0];
+        $image_path = str_replace('/'.$image_name, '', $source);
+        File::deleteDirectory($source);
+        $image->move($image_path, $image_name);
+        return $source;
+//        $ratio = (int)$request->naturalWidth / (int)$request->width;
+//        $ratio = (int)$ratio;
+//        $width = (int)$request->width;
+//        $height =  (int)$request->height;
+//        $left = (int)$request->left * $ratio;
+//        $top = (int)$request->top * $ratio;
+//
+//        $source = str_replace(URL::to('/') . '/', '', $request->source);
+//        $img = Image::make($source);
+//        $img->crop($width, $height, $left, $top);
+//        $img->save($source);
+
     }
     public function adminPostUpdateCurrencies(Request $request)
     {
